refactor: Extract form hook & hide save button with flag
